+++ src/main/java/com/tanjiaming99/community/dto/QuestionDTO.java
@@ -15,13 +15,13 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private  String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer Creator;
+    private Long Creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
